#6 use webpack to build extension
