+++ package.json
@@ -1,0 +1,28 @@
+{
+  "name": "Luca",
+  "version": "1.0.0",
+  "description": "A chrome extension to watch movies together",
+  "main": "index.js",
+  "scripts": {
+    "build": "webpack --config webpack.config.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/abozanona/luca.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/abozanona/luca/issues"
+  },
+  "homepage": "https://github.com/abozanona/luca#readme",
+  "devDependencies": {
+    "@types/chrome": "0.0.183",
+    "copy-webpack-plugin": "^9.1.0",
+    "ts-loader": "^9.2.9",
+    "typescript": "^4.6.4",
+    "webpack": "^5.72.0",
+    "webpack-cli": "^4.9.2"
+  }
+}
