+++ src/js/background.ts
