+++ src/js/popup.ts
@@ -1,13 +1,13 @@
 chrome.runtime.connect({ name: "popup" });
 
-function getCurrentTabId(cb) {
+function getCurrentTabId(cb: any) {
 	chrome.tabs.query({ active: true, currentWindow: true }, function (tabs) {
 		cb(tabs[0].id);
 	});
 }
 
 document.getElementById("btn-highlight-page-videos").addEventListener("click", function (event) {
-	getCurrentTabId(function (tabId) {
+	getCurrentTabId(function (tabId: any) {
 		const message = {
 			code: 'Q_HIGHLIGHT_ALL_VIDEOS',
 		}
@@ -16,7 +16,7 @@ document.getElementById("btn-highlight-page-videos").addEventListener("click", f
 });
 
 document.getElementById("btn-create-new-room").addEventListener("click", function (event) {
-	getCurrentTabId(function (tabId) {
+	getCurrentTabId(function (tabId: any) {
 		const message = {
 			code: 'Q_CREATE_NEW_ROOM_ID',
 		}
@@ -25,10 +25,10 @@ document.getElementById("btn-create-new-room").addEventListener("click", functio
 });
 
 document.getElementById("btn-join-room").addEventListener("click", function (event) {
-	getCurrentTabId(function (tabId) {
+	getCurrentTabId(function (tabId: any) {
 		const message = {
 			code: 'Q_JOIN_ROOM_ID',
-			roomId: document.getElementById("txt-room-id").value,
+			roomId: (document.getElementById("txt-room-id") as HTMLInputElement).value,
 		}
 		chrome.tabs.sendMessage(tabId, message);
 	});
@@ -37,8 +37,8 @@ document.getElementById("btn-join-room").addEventListener("click", function (eve
 Array.from(document.getElementsByClassName("img-avatar")).forEach(element => {
 	element.addEventListener("click", function (el) {
 		document.getElementsByClassName("img-avatar-selected")[0].classList.remove('img-avatar-selected');
-		el.target.classList.add('img-avatar-selected');
-		let avatarName = el.target.dataset.name;
+		(el.target as HTMLElement).classList.add('img-avatar-selected');
+		let avatarName = (el.target as HTMLElement).dataset.name;
 		chrome.storage.sync.set({ userAvatar: avatarName }, function () {
 
 		});
@@ -57,7 +57,7 @@ chrome.storage.sync.get('userAvatar', function (items) {
 });
 
 document.getElementById("txt-user-name").addEventListener("change", function (el) {
-	chrome.storage.sync.set({ userName: el.target.value }, function () {
+	chrome.storage.sync.set({ userName: (el.target as HTMLInputElement).value }, function () {
 
 	});
 });
@@ -66,20 +66,17 @@ chrome.storage.sync.get('userName', function (items) {
 	let userName = items.userName;
 	if (!userName) {
 		var randomNames = ['Luca Paguro', 'Alberto Scorfano', 'Giulia Marcovaldo', 'Ercole Visconti', 'Daniela Paguro', 'Massimo Marcovaldo', 'Lorenzo Paguro', 'Ciccio', 'Guido', 'Signora Marsigliese', 'Tommaso', 'Grandma Paguro', 'Giacomo', 'Concetta Aragosta', 'Pinuccia Aragosta', 'Uncle Ugo', 'Maggiore'];
-		randomNames.random = function () {
-			return this[Math.floor(Math.random() * this.length)];
-		};
 
-		userName = randomNames.random();
+		userName = randomNames[Math.floor(Math.random() * randomNames.length)];
 		chrome.storage.sync.set({ userName: userName }, function () {
 
 		});
 	}
-	document.getElementById("txt-user-name").value = userName;
+	(document.getElementById("txt-user-name") as HTMLInputElement).value = userName;
 });
 
 function getCurrentPageStatus() {
-	getCurrentTabId(function (tabId) {
+	getCurrentTabId(function (tabId: any) {
 		const message = {
 			code: 'Q_GET_PAGE_STATUS'
 		}
