+++ tsconfig.json
@@ -1,0 +1,16 @@
+{
+    "compilerOptions": {
+        "strict": true,
+        "module": "commonjs",
+        "target": "es6",
+        "esModuleInterop": true,
+        "sourceMap": true,
+        "rootDir": "src",
+        "outDir": "dist/js",
+        "noEmitOnError": true,
+        "typeRoots": [
+            "node_modules/@types"
+        ],
+        "strictNullChecks": false
+    }
+}
